HTML5 validation, removal of unused css.
Week1
=====

index.html
----------

*fixed spelling errors(codercojo)
*fixed capitalization(facebook, Iphone)
*added/removed some white space
*added lang attribute to html for WCAG validation
*added an alt attribute to img for WCAG and validation purposes
*added charset meta tag for validation purposes
*added doctype for html5 removed closing tag for img to meet html5
validation

Week2
=====

index.html
----------

*fixed capitalization(coderdojo iphone)
*changed code layout to make it valid(you cannot have a un-order list
inside a paragraph)
*added/removed some white space
*added lang attribute to html for WCAG reasons
*added charset meta tag for validation purposes
*added doctype for html5 removed closing tag for img and link to meet
html5 validation
*removed iframe frameborder attribute, all styling should be done via
css

style.css
---------
*changed font-family because comic sans ms is not a websafe font(works
on win, osx but not Unix+X) and it is a "uncool" font.
*added iframe styling to remove borders

Week3
=====

Are we doing xhtml 4.01 transitional, xhtml 4.01 strict, html 4.01
strict, html 4.01 transitional or html5?
If html5(What I believe we should be doing) only <!DOCTYPE html> is
needed.

Did not update the readme file to take care of these changes

index.html
----------
*changed doctype to <!DOCTYPE html>. If the old doctype was kept we
were doing xhtml 4.01 traditional. that said the old doctype line had
errors, it ended with />, you end a doctype with a single closing angle
bracket >. If xhtml 4.01 traditional was being used the xmlns attribute
for html would need to be set.
*added lang attribute to html for WCAG reasons
*added charset meta tag for validation purposes
*fixed capitalization(coderdojo iphone)
*changed code layout to make it valid(you cannot have a un-order list
inside a paragraph)
*changed div class to id and renamed it to container. Renamed to
container as it is the site container i.e. it sets the width and
position of all site content. changed from class to id as it can/should
only occur once.
*removed ./ from the link for the style sheet, its not required.
*centered the youtube video to also show centering and wrapping of divs

style.css
----------
*removed div.logoDiv and img as they are not used.
*changed div.mainDiv to suit the my change of a class to id and
renaming of it as described in index.html
*changed font-family because comic sans ms is not a websafe font(works
on win, osx but not Unix+X) and it is a "uncool" font.
*added center class to using divs as wrapping in html